move spinnerlogo to loaders
