+++ src/components/loading/Loading.stories.mdx
@@ -2,6 +2,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
 import { colors } from 'common/colors';
 import Spinner from './Spinner';
+import SpinnerLogo from './SpinnerLogo';
 
 <Meta title="Feedback|Spinner" component={Spinner} />
 
@@ -19,3 +20,29 @@ import Spinner from './Spinner';
 </Preview>
 
 <Props of={Spinner} />
+
+# Spinner Logo
+
+<Preview>
+  <Story name="Spinner Logo">
+    <>
+      <h3>Classic Spinner</h3>
+      <div style={{ display: 'inline-flex' }}>
+        <SpinnerLogo />
+      </div>
+      <h3>Spinner with other colors</h3>
+      <div style={{ display: 'inline-flex' }}>
+        <SpinnerLogo primaryColor={'#db7093'} secondaryColor={'#2f9c0a'} />
+      </div>
+      <h3>Spinners of different sizes</h3>
+      <div style={{ display: 'inline-flex' }}>
+        <SpinnerLogo spinnerSize={'75px'} />
+        <SpinnerLogo spinnerSize={'100px'} />
+        <SpinnerLogo spinnerSize={'150px'} />
+      </div>
+    </>
+  </Story>
+</Preview>
+
+<Props of={SpinnerLogo} />
+
