+++ src/components/logo/Logo.stories.mdx
@@ -35,9 +35,7 @@ import Logo from './Logo';
 
 <Props of={StaticLogo} />
 
-import SpinnerLogo from './SpinnerLogo';
-
-# Spinner Logo
+# Static Spinner
 
 <Preview>
   <Story name="Spinner Logo">
