+++ src/index.ts
@@ -13,7 +13,7 @@ export { default as EventCard } from './components/eventCard/EventCard';
 export { default as CircleCross } from 'components/icons/CircleCross';
 export { default as CircleCheckmark } from 'components/icons/CircleCheckmark';
 export { default as Spinner } from 'components/loading/Spinner';
-export { default as SpinnerLogo } from 'components/logo/SpinnerLogo';
+export { default as SpinnerLogo } from 'components/loading/SpinnerLogo';
 export { default as StaticLogo } from 'components/logo/StaticLogo';
 export { default as TextField } from 'components/textField/TextField';
 export { default as ToggleSwitch } from 'components/toggleSwitch/ToggleSwitch';
